requirements.txt
spotipy
google.cloud.storage
pandas
google.cloud.bigquery